requirements.txt
cython
pillow